Correction of map representation in RESP2 (#2587)
In the protocol spec, the description of how maps are represented in RESP2 is incorrect.

"Maps in RESP2 are represented by arrays, in which each element is a key-value tuple. Each tuple is an array with two elements, these being the key and the value."

This is not true. A map in RESP2 is represented as a flat array containing the keys and values, not nested arrays. The description is updated.

Co-authored-by: Viktor Söderqvist <viktor.soderqvist@est.tech>